+++ src/api/user/user_service.go
@@ -40,7 +40,7 @@ func (s *Service) GenerateSalt(ctx context.Context, saltNonce, email string) str
 	return b64.StdEncoding.EncodeToString(h.Sum(nil))
 }
 
-func (s *Service) CreateUser(ctx context.Context, username, email, password string) (int, error) {
+func (s *Service) CreateUser(ctx context.Context, username, email, password string) (int64, error) {
 	salt := s.GenerateSalt(context.Background(), s.configuration.Opaque.SaltNonce, email)
 
 	id, err := s.db.InsertUser(username, email, password, salt)
