+++ src/models/api_key.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type ApiKey struct {
-	ID        int        `json:"id"`
-	UserID    int        `json:"user_id"`
+	ID        int64      `json:"id"`
+	UserID    int64      `json:"user_id"`
 	Name      string     `json:"name"`
 	Key       string     `json:"key"`
 	CreatedAt time.Time  `json:"created_at"`
